Вывод гистограммы

+++ ComputerModelling/ComputerModelling/CoveyouGenerator.h
@@ -8,6 +8,6 @@ private:
 public:
 	CoveyouGenerator(long initial_seed, short degree_of_two);
 
-	long next();
+	double next();
 };
